feat: latest uncommitted update

+++ Controllers/JournalEntryController.cs
@@ -1,7 +1,9 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
+using Microsoft.EntityFrameworkCore;
 using WebAPI.DAL;
+using WebAPI.Data.Enum;
+using WebAPI.Exception;
 using WebAPI.Models;
 using WebAPI.Models.DTO;
 using WebAPI.Models.Mapper;
@@ -41,6 +43,11 @@ namespace WebAPI.Controllers
         public async Task<ActionResult<JournalEntry>> PostJournalEntry([FromBody] JournalEntryDto journalEntryDto)
         {
             JournalEntry journalEntry = JournalEntryMapper.JournalEntryDtoToJournalEntry(journalEntryDto);
+
+            bool closingEntryExist = await _journalEntryRepository.ClosingEntryExist(journalEntry.TransactionDate);
+            if (closingEntryExist)
+                throw new DatabaseInsertException("Periode yang terpilih sudah ditutup, tidak dapat melakukan penambahan data.");
+
             await _journalEntryRepository.InsertJournalEntry(journalEntry);
 
             return CreatedAtAction(nameof(GetJournalEntryById), new { id = journalEntry.Id }, journalEntry);
@@ -57,9 +64,16 @@ namespace WebAPI.Controllers
                 return BadRequest($"Entri jurnal dengan id: {id} tidak ditemukan");
 
             journalEntryUpdateDto.PassData(ref journalEntry);
+
+            bool closingEntryExist = await _journalEntryRepository.ClosingEntryExist(journalEntry.TransactionDate);
+            if (closingEntryExist)
+                throw new DatabaseUpdateException("Periode yang terpilih sudah ditutup, tidak dapat melakukan perubahan data.");
+
             await _journalEntryRepository.UpdateJournalEntry(journalEntry);
 
-            return CreatedAtAction(nameof(GetJournalEntryById), new { id = journalEntry.Id }, journalEntry);
+            journalEntry = await _journalEntryRepository.GetJournalEntryById(id);
+
+            return CreatedAtAction(nameof(GetJournalEntryById), new { id = journalEntry!.Id }, journalEntry!);
         }
 
         [HttpDelete("delete/{id}")]
@@ -69,9 +83,30 @@ namespace WebAPI.Controllers
             if (journalEntry is null)
                 return BadRequest($"Data jurnal dengan id: {id} tidak ditemukan!");
 
+            bool closingEntryExist = await _journalEntryRepository.ClosingEntryExist(journalEntry.TransactionDate);
+            if (closingEntryExist)
+                throw new DatabaseDeleteException("Periode yang terpilih sudah ditutup, tidak dapat menghapus entri.");
+
+            if (journalEntry.Status == EntryStatus.Posted.ToString())
+                throw new DatabaseDeleteException("Tidak dapat menghapus entri yang sudah di posting!");
+
             await _journalEntryRepository.DeleteJournalEntry(journalEntry);
 
             return Ok();
+        }
+
+        [HttpGet("purchase/{purchaseHeaderId}")]
+        public async Task<ActionResult<PurchaseEntryContainer>> GetPurchaseReportById(Guid purchaseHeaderId)
+        {
+            IEnumerable<PurchaseEntryContainer> purchaseReportData = await _journalEntryRepository.GetPurchaseQueryById(purchaseHeaderId);
+            return Ok(purchaseReportData);
+        }
+
+        [HttpGet("sales/{salesHeaderId}")]
+        public async Task<ActionResult<SalesEntryContainer>> GetSalesReportById(Guid salesHeaderId)
+        {
+            IEnumerable<SalesEntryContainer> salesReportData = await _journalEntryRepository.GetSalesQueryById(salesHeaderId);
+            return Ok(salesReportData);
         }
